Refactored AppConfig code
